refactor how the retry works
I did this pretty quick the other day, and after thinking about it I
wasn't happy with the way the retry worked. Added a helper to make
retry's better and not rely on the built in retry keyworkd. I've also
added tests to make sure that the flow works the way as expected.